requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
gunicorn>=21.2
pandas>=2.1
openpyxl>=3.1
python-dotenv>=1.0
